requirements.txt
Flask
python-dotenv
psycopg2-binary 
PyPDF2
pytesseract
pdf2image
Pillow 